Renamed rejection-sagas.spec.js to rejection-sagas.test.js and cleaned up the rejection sagas tests.
