+++ src/test.js
@@ -1,5 +1,5 @@
 import './features/rejection/rejection-reducer.test';
-import './features/rejection/rejection-sagas.spec';
+import './features/rejection/rejection-sagas.test';
 import './features/rejection/add-question/add-question-component.test';
 import './features/rejection/question-list/question-list-component.test';
 import './features/rejection/question/question-component.test';
